Source Monday: fix schema loader; delete old files
