+++ airbyte-integrations/connectors/source-monday/source_monday/__init__.py
@@ -2,11 +2,8 @@
 # Copyright (c) 2021 Airbyte, Inc., all rights reserved.
 #
 
-
 from .dpath_string_extractor import DpathStringExtractor
 from .graphql_request_options_provider import GraphQLRequestOptionsProvider
-
-# from .source import SourceMonday
-from .source_lc import SourceMonday
+from .source import SourceMonday
 
 __all__ = ["SourceMonday", "GraphQLRequestOptionsProvider", "DpathStringExtractor"]
