+++ airbyte-integrations/connectors/source-monday/source_monday/monday.yaml
@@ -2,7 +2,7 @@ version: "0.1.0"
 
 definitions:
   schema_loader:
-    type: JsonSchema
+    type: JsonFileSchemaLoader
     file_path: "./source_monday/schemas/{{ options['name'] }}.json"
   selector:
     type: RecordSelector
@@ -18,6 +18,7 @@ definitions:
       type: BearerAuthenticator
       api_token: "{{ config['credentials']['api_token'] if config['credentials']['auth_type'] == 'api_token' else config['credentials']['access_token'] if config['credentials']['auth_type'] == 'oauth2.0' else config.get('api_token', '') }}"
     request_options_provider:
+      type: InterpolatedRequestOptionsProvider
       class_name: "source_monday.GraphQLRequestOptionsProvider"
       limit: "{{ options['items_per_page'] }}"
     error_handler:
@@ -59,7 +60,7 @@ definitions:
       record_selector:
         type: RecordSelector
         extractor:
-          type: DpathExtractor
+          type: CustomRecordExtractor
           field_pointer: "/data/boards/*/items/*"
           class_name: "source_monday.DpathStringExtractor"
       paginator:
