--- airbyte-integrations/connectors/source-monday/source_monday/source_lc.py
@@ -1,17 +1,0 @@
-#
-# Copyright (c) 2022 Airbyte, Inc., all rights reserved.
-#
-
-from airbyte_cdk.sources.declarative.yaml_declarative_source import YamlDeclarativeSource
-
-"""
-This file provides the necessary constructs to interpret a provided declarative YAML configuration file into
-source connector.
-WARNING: Do not modify this file.
-"""
-
-
-# Declarative Source
-class SourceMonday(YamlDeclarativeSource):
-    def __init__(self):
-        super().__init__(**{"path_to_yaml": "monday.yaml"})
